java/client/test/org/openqa/selenium/support/pagefactory/ByCompositeTest.java
// Licensed to the Software Freedom Conservancy (SFC) under one
// or more contributor license agreements.  See the NOTICE file
// distributed with this work for additional information
// regarding copyright ownership.  The SFC licenses this file
// to you under the Apache License, Version 2.0 (the
// "License"); you may not use this file except in compliance
// with the License.  You may obtain a copy of the License at
//
//   http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing,
// software distributed under the License is distributed on an
// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
// KIND, either express or implied.  See the License for the
// specific language governing permissions and limitations
// under the License.

package org.openqa.selenium.support.pagefactory;

import static org.hamcrest.Matchers.equalTo;
import static org.junit.Assert.assertThat;
import static org.junit.Assert.fail;
import static org.mockito.Mockito.mock;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.junit.runners.JUnit4;
import org.openqa.selenium.By;
import org.openqa.selenium.NoSuchElementException;
import org.openqa.selenium.SearchContext;
import org.openqa.selenium.WebElement;
import org.openqa.selenium.internal.FindsById;
import org.openqa.selenium.internal.FindsByLinkText;
import org.openqa.selenium.internal.FindsByName;
import org.openqa.selenium.internal.FindsByXPath;

import java.util.Collections;
import java.util.List;

@RunWith(JUnit4.class)
public class ByCompositeTest {

  private ByComposite mockByComposite(By... bys) {
    return new ByComposite(bys) {
      @Override
      public String getOperation() {
        return "mock";
      }

      @Override
      public List<WebElement> findElements(SearchContext context) {
        return Collections.emptyList();
      }
    };
  }

  @Test
  public void findElementZeroBy() {
    final AllDriver driver = mock(AllDriver.class);
    ByComposite by = mockByComposite();
    try {
      by.findElement(driver);
      fail("Expected NoSuchElementException!");
    } catch (NoSuchElementException expected) {
      // Expected
    }
  }

  @Test
  public void testEquals() {
    assertThat(mockByComposite(By.id("cheese"), By.name("photo")),
        equalTo(mockByComposite(By.id("cheese"), By.name("photo"))));
  }

  private interface AllDriver extends
      FindsById, FindsByLinkText, FindsByName, FindsByXPath, SearchContext {
    // Place holder
  }

}
